#86c039xju - refactor cart methods

+++ src/main/java/org/sluja/searcher/webapp/service/user/cart/ICart.java
@@ -1,13 +1,11 @@
 package org.sluja.searcher.webapp.service.user.cart;
 
-import jakarta.servlet.http.HttpSession;
-
 import java.util.List;
 
 public interface ICart<R, T> {
     void addProductToCart(T product);
     void removeProductFromCart(R product);
     void clearCart();
-    void changeQuantity(R product, Long quantity);
+    void changeQuantity(R product, int quantity);
     List<R> getCartProducts();
 }
